Changes to the CldOvlp_2018VP merge into HWRF, as a result of the Consistency Check.
When running the Consistency Check we found the operational HWRF
answers had slightly changed when the trunk is run with default
settings (cldovrlp=4 and idcor=0).

This difference was a result of changing the decorrelation
length to the inverse and multiplying instead of keeping
it as division.

was changed to this:
alpha(i, ilev) = exp( -(hgt(i,ilev) - hgt(i,ilev-1)) * Zo_inv(i))

is now this:
alpha(i, ilev) = exp( -(hgt(i,ilev) - hgt(i,ilev-1)) / Zo(i))

Mathematically the two equations are equivalent, but results in
rounding/precison differences when computing.

This change reverts back to dividing and the CC now passes.