MYNN PBL restart differences: add restart output stream for el_pbl (#971)
TYPE: bug fix

KEYWORDS: bug, MYNN, registry, el_pbl, restart, differences

SOURCE: internal

DESCRIPTION OF CHANGES: When using the MYNN PBL scheme, restarts were not giving
bit-for-bit results when comparing the output of a standard run with that of a restart run. This
was thought to have been corrected in d944f1e, but was not. When using MYNN, the default
setting for icloud_bl is 1 (on), but a variable that is used (el_pbl) was not put in the restart stream.
Simply adding this variable to the restart stream in Registry.EM_COMMON corrects the problem.

LIST OF MODIFIED FILES:
M Registry/Registry.EM_COMMON

TESTS CONDUCTED:
Confirmed that the code builds, and the change corrects the non-bit-for-bit restart problem.

RELEASE NOTE: When using the MYNN PBL scheme, with icloud_bl = 1 (which is default), restarts did not give bit-for-bit results when compared to a non-restart run. This has been a problem since the option was introduced in V3.8, but is now corrected.