Changes to `arch/` directory to support Cygwin.
Changed treatment of preamble parsing to make it more similar to
postamble parsing.  It would probably be a good idea to unify the
treatment of the three sections.  The other option is to change the
condition on the #NOWIN section in the postamble, but I'm not sure
that works as intended.

This way the default is specified before overrides, which is nice.