H220baseline: Bug fixes for the Ferrier-Aligo microphysics scheme from Eric Aligo.
a. The CIACW related bug fix
CIACW is a constant used for calculating collection of cloud water by large
ice.  In 2015, it was found there was a double counting of the air resistance
factor.  CIACW bug was corrected in the NAM and HMON.  CIACW fix was put in
HWRF in SUBROUTINE fer_hires_init, but bug remains in  HWRF on line 329:
CIACW=DT*0.25*PI_E*0.5*(1.E5)**C1.
Correct calculation should be:
CIACW=0.5*DTPH*0.25*PI_E

b. The QSW0 related fix
QSW0 is the saturation mixing ratio w.r.t. water at 0C.  Calculation modified
to make it more robust.  Prevents the very low chance that ice may form in the
stratosphere when T> 0C and P<7 hPa during sudden stratospheric warming events.
These events without the change led to NAM failures.  Forecast not changed (bit
identical).  Change was made in the NAM, but not in HMON or HWRF.