NMM needs Registry mods for new var in bdy_control
The lateral BC option to allow multiple files (one time period per
file) needs to be in the NMM Registry also. This option is likely
to not be used by NMM, but the WRF code does not compile without
this variable being declared as part of the namelist config_flags.

Optionally, an EM_ONLY ifdef could have been put around the code to
protect it during an NMM build.

modified:   Registry/Registry.NMM