Remove unused irr_chan from lsm and lsm_mosaic
Problem:
The irr_chan variable is declared OPTIONAL, and is not passed into the
lsm routine or the lsm_mosaic routine. In both routines the array
irr_chan is used without an IF(PRESENT(irr_chan)). This causes a
seg fault when running the Noah LSM scheme.

The purpose of the irr_chan array is to fill in a local real value,
irrigation_channel. This variable is never used inside of either
lsm or lsm_mosaic.

Solution:
Remove the irr_chan and irrigation_channel from both the subroutines
lsm and lsm_mosaic.

modified:   module_sf_noahdrv.F