Fix feedback for odd nest ratios > 3:1
TYPE: bug fix

KEYWORDS: nest feedback

SOURCE: Reported by Jamie Bresch and Kevin Manning (NCAR), fixed internally

DESCRIPTION OF CHANGES:
A fixed constant value of 1 was correct for a 3:1 ratio since the term was
supposed to be ratio/2. For a 5:1 ratio the value needed to be 2. Replacing
the fixed value with the correct term fixes the problem.

LIST OF MODIFIED FILES:
modified:   share/interp_fcn.F

TESTS CONDUCTED:

The test involves generating a known block of values on the fine grid that
should all feedback to their respective parent grid cells. By choosing the
fine grid cells so that only full parent cells are selected, and by filling
in the fine grid with a constant value (in our case: 1.0), then the coarse
grid should reflect the fine grid after feedback.

1. Example of wrong indexing.

2. Example of simple assignment.

3. Example with fixed but now standard code.

RELEASE NOTE: The odd nest ratios (excluding 3:1) had an incorrect feedback. The indexing was off by (ratio/2 -1), so off by a single fine grid point for a 5:1 ratio.