Allow DA and FSBM to peacefully coexist in same repo
TYPE: bug fix

KEYWORDS: DA FSBM real*8

SOURCE: internal

DESCRIPTION OF CHANGES:
The Fast Spectral Bin Microphysics (FSBM) code only works with default real
size = 4 bytes due to a mixture of declarations using real, real*8,
and double precision. To solve this problem, the FSBM code is ifdef'ed out
for any real*8 build. This is handled via the configure step.

Problem:
The DA code is built with real*8 by default without resorting to any special
setup in the configure file. The choice was to either disable FSBM, or not
build DA.

Solution:
This PR identifies the DA builds, and sets internal flags to let the build
know to ifdef out the FSBM code. If no DA code is requested AND if this is
not a real*8 build, then the FSBM code is part of the build executable.

LIST OF MODIFIED FILES:
M	arch/postamble
M	configure

TESTS CONDUCTED:
1. Jenkins tests real*4 and real*8 builds.
2. DA group will test the various DA builds.