syntax error in irrigation
TYPE: bug fix

KEYWORDS: irrigation, drip, statement function

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
In module_irrigation.F, in subroutine sprinkler_irrigation is this line:
```
constants_irrigation=irr_freq*irr_daily_amount/(irr_num_hours*3600*rho(a,kms,b)*dz8w(a,kms,b)*100)
```
It is clear from the rest of the routine, that this is to be treated as a
statement function.
1. The line is after the declarations and before the first executable statement.
2. The values of a, b, and kms are not yet defined.
3. The following line _tries_ to use it as a statement function.
```
qr_curr(a,kms,b)=qr_curr(a,kms,b)+irrigation(a,b)*constants_irrigation*dt
```

However, the GNU compiler is segfaulting on the definition of the statement function,
because the variables "a" and "b" (among others) are not yet assigned and are likely
garbage values. It is segfaulting because it does not recognize this as a proper statement
function.

Statement functions were declared obsolescent in f95, but we have them in other routines.

Solution:
Abandon the statement function completely.

LIST OF MODIFIED FILES:
modified:   module_irrigation.F

TESTS CONDUCTED:
1. With no mod, with
```
 &physics
 sf_surf_irr_scheme                  = 3,     3,     3,
```
We get:
```
WRF NUMBER OF TILES =   1

Program received signal SIGSEGV: Segmentation fault - invalid memory reference.

Backtrace for this error:
	at /wrf/WRF/phys/module_irrigation.f90:140
```
2. With the mods, the code runs to completion.