FSBM: use existing package flag for T/F compute diagnostics
TYPE: bug fix

KEYWORDS: FSBM, IPolar_HUCM, sbm_diagnostics, seg fault

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
The IPolar_HUCM variable is a LOGICAL PARAMETER that is always TRUE. It is a
module local variable. When we get down to the IF test in FSBM to determine if
we are doing diagnostics, we therefore are always proceeding with diagnostic
calculations. If the sbm_diagnostics flag is zero, then the sbmradar array is
only "unit" allocated. We get a segfault when filling the array sbmradar.

Solution:
The sbm_diagnostics namelist variable is used in the registry.polrad to package
the sbmradar array. Use that same variable to determine if we compute the
diagnostics.

LIST OF MODIFIED FILES: list of changed files (use `git diff --name-status master` to get formatted list)
modified:   dyn_em/solve_em.F
modified:   phys/module_microphysics_driver.F
modified:   phys/module_mp_fast_sbm.F

TESTS CONDUCTED:
 - [x] Without mod, seg faults
 - [x] With mod, code runs to completion