If WRF model has sf_urban_physics==0, real namelist value is ignored
TYPE: enhancement

KEYWORDS: sf_urban_physics

SOURCE: internal

DESCRIPTION OF CHANGES:
Problem:
If a user runs the real program with an urban physics option activated, then the
WRF model needs that same option activated. However, if the WRF model just turns
off the urban option, then ANY setting of the real program's namelist value of
sf_urban_physics should be acceptable.

Solution:
Add an extra IF test to check on the WRF model's namelist setting for sf_urban_physics. If
this namelist option is non-zero, then there is a validation test to determine if the
namelist and the input file metadata are identical.

This is consistent with the treatment of several other flags that are set by the real
program, for example: gwd_opt and sf_ocean_physics.

LIST OF MODIFIED FILES:
modified:   share/input_wrf.F

TESTS CONDUCTED:
1. Reggie
2. 2x2 contingency table for sf_urban_physics consistency

|                | NML=0 | NML=1 | NML=2 | NML=3 |
|----------------|:-----:|:-----:|:-----:|:-----:|
|File Metadata=0 |       |       |       |       |
|File Metadata=1 |       |       |       |       |
|File Metadata=2 |       |       |       |       |
|File Metadata=3 |       |       |       |       |