Efficiency fix for nwp_diag reflectivity calculation (#1170)
TYPE:  enhancement

KEYWORDS: nwp_diagnostics, diag_flag

SOURCE: Ted Mansell (NSSL)

DESCRIPTION OF CHANGES: 
Replaces "Efficiency fix for nwp_diag reflectivity calculation" #1060

Problem:
With commit 86f7675359234d239, PR #749 "make refl_10cm available if nwp_diagnostics is on", 
solve_em.F was changed so that `diag_flag` was always set to true when `nwp_diagnostics==1`. 
This caused the full 3D radar reflectivity to be calculated on each time step, whereas the diagnostic 
for the max dBZ only needs to compute lowest level (k=1). The full 3D nature of the new computation
makes the cost of this diagnostic between one and two orders of magnitude more expensive than 
necessary. 

Solution:
A variable `ke_diag` is added in solve_em to tell the reflectivity calculation whether to do the full 
depth (for history output) or only the lowest level (for `nwp_diagnostics==1`). This avoids 
unnecessary 3Dcomputation of dBZ for `nwp_diagnostics==1`, which is expensive for a number 
of schemes. The variable `ke_diag` is passed to the microphysics driver and from there to individual 
microphysics schemes. The NSSL and Thompson schemes have been updated to use this new 
capability, and other MP schemes could be done in a very similar manner.

LIST OF MODIFIED FILES: 
 dyn_em/solve_em.F
 phys/module_microphysics_driver.F
 phys/module_mp_nssl_2mom.F
 phys/module_mp_thompson.F

TESTS CONDUCTED: 
1. Ideal full-physics tests showed that the NWP field REFD_MAX is the same after implementation.
2. Jenkins all PASS.

RELEASE NOTE: Improved efficiency of reflectivity calculation for nwp_diagnostics=1 by only computing it at the level where maximum reflectivity is diagnosed. Currently this performance fix is only available in the NSSL 2-moment and Thompson MP schemes.