Add C1 and C2 arrays as args to advection drivers
With the introduction of the hybrid coordinate, two additional 1d arrays
need to be passed in as args to the advection drivers.

c1 = d(B)/d(eta)
c2 = (1-c1)*(p00-ptop)

For this simplified test, c1=1 and c2=0.

Also, an oops was fixed in the title for the WENO PD for gnuplot.

modified:   module_advect_em.F