fixed slope calculation at domain boundary for periodic boundary conditions (#1217)
TYPE: enhancement

KEYWORDS: slope, periodic boundary conditions

SOURCE: Matthias Göbel (University of Innsbruck)

DESCRIPTION OF CHANGES: In the current WRF version the slope angle on the lateral boundaries of the domain is calculated differently than in the interior of the domain. For the former, forward or backward final differences are used, for the latter central finite differences. This can lead to different shortwave fluxes for equivalent locations in mountainous terrain.
For periodic boundary conditions, the terrain outside of the domain is known and set right before.
This enhancement implements central differences also for the domain boundary to get the same shortwave flux for equivalent locations.

ISSUE: Fixes #1008

LIST OF MODIFIED FILES: dyn_em/start_em.F

TESTS CONDUCTED:

RELEASE NOTE: for periodic boundary conditions the slope calculation for the lateral boundaries is made equivalent to the calculation for the interior of the domain