Promote wwE and wwI to "i1" variables for performance
The memory-sized local variable wwE and wwI were allocated on the stack
for rk_tendency, and for each of the multiple calls to rk_scalar_tend.
These are each called for every RK time step. The mods allocate the space
in the solver once per model time step, and since the variables are "i1",
they are not kept during a nested domain.

 Changes to be committed:
	modified:   Registry/Registry.EM_COMMON
	modified:   dyn_em/module_em.F
	modified:   dyn_em/solve_em.F